# WIP - tsvector columns with gin