--- a/config/application.rb
+++ b/config/application.rb
@@ -30,5 +30,6 @@ module AllSorns
     # the framework and any gems in your application.
 
     config.active_job.queue_adapter = :good_job
+    config.active_record.schema_format = :sql
   end
 end