--- a/db/schema.rb
+++ b/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_12_28_232801) do
+ActiveRecord::Schema.define(version: 2020_12_30_173611) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -34,7 +34,7 @@ ActiveRecord::Schema.define(version: 2020_12_28_232801) do
     t.index ["sorn_id"], name: "index_agencies_sorns_on_sorn_id"
   end
 
-  create_table "good_jobs", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "good_jobs", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.text "queue_name"
     t.integer "priority"
     t.jsonb "serialized_params"
@@ -96,6 +96,33 @@ ActiveRecord::Schema.define(version: 2020_12_28_232801) do
     t.string "publication_date"
     t.string "title"
     t.string "action_type"
+    t.tsvector "agency_names_vector"
+    t.tsvector "action_vector"
+    t.tsvector "summary_vector"
+    t.tsvector "dates_vector"
+    t.tsvector "addresses_vector"
+    t.tsvector "further_info_vector"
+    t.tsvector "supplementary_info_vector"
+    t.tsvector "system_name_vector"
+    t.tsvector "system_number_vector"
+    t.tsvector "security_vector"
+    t.tsvector "location_vector"
+    t.tsvector "manager_vector"
+    t.tsvector "authority_vector"
+    t.tsvector "purpose_vector"
+    t.tsvector "categories_of_individuals_vector"
+    t.tsvector "categories_of_record_vector"
+    t.tsvector "source_vector"
+    t.tsvector "routine_uses_vector"
+    t.tsvector "storage_vector"
+    t.tsvector "retrieval_vector"
+    t.tsvector "retention_vector"
+    t.tsvector "safeguards_vector"
+    t.tsvector "access_vector"
+    t.tsvector "contesting_vector"
+    t.tsvector "notification_vector"
+    t.tsvector "exemptions_vector"
+    t.tsvector "history_vector"
     t.index "to_tsvector('english'::regconfig, (access)::text)", name: "access_idx", using: :gist
     t.index "to_tsvector('english'::regconfig, (action)::text)", name: "action_idx", using: :gist
     t.index "to_tsvector('english'::regconfig, (addresses)::text)", name: "addresses_idx", using: :gist
@@ -123,7 +150,34 @@ ActiveRecord::Schema.define(version: 2020_12_28_232801) do
     t.index "to_tsvector('english'::regconfig, (supplementary_info)::text)", name: "supplementary_info_idx", using: :gist
     t.index "to_tsvector('english'::regconfig, (system_name)::text)", name: "system_name_idx", using: :gist
     t.index "to_tsvector('english'::regconfig, (system_number)::text)", name: "system_number_idx", using: :gist
+    t.index ["access_vector"], name: "access_vector_idx", using: :gin
+    t.index ["action_vector"], name: "action_vector_idx", using: :gin
+    t.index ["addresses_vector"], name: "addresses_vector_idx", using: :gin
+    t.index ["agency_names_vector"], name: "agency_names_vector_idx", using: :gin
+    t.index ["authority_vector"], name: "authority_vector_idx", using: :gin
+    t.index ["categories_of_individuals_vector"], name: "categories_of_individuals_vector_idx", using: :gin
+    t.index ["categories_of_record_vector"], name: "categories_of_record_vector_idx", using: :gin
     t.index ["citation"], name: "index_sorns_on_citation"
+    t.index ["contesting_vector"], name: "contesting_vector_idx", using: :gin
+    t.index ["dates_vector"], name: "dates_vector_idx", using: :gin
+    t.index ["exemptions_vector"], name: "exemptions_vector_idx", using: :gin
+    t.index ["further_info_vector"], name: "further_info_vector_idx", using: :gin
+    t.index ["history_vector"], name: "history_vector_idx", using: :gin
+    t.index ["location_vector"], name: "location_vector_idx", using: :gin
+    t.index ["manager_vector"], name: "manager_vector_idx", using: :gin
+    t.index ["notification_vector"], name: "notification_vector_idx", using: :gin
+    t.index ["purpose_vector"], name: "purpose_vector_idx", using: :gin
+    t.index ["retention_vector"], name: "retention_vector_idx", using: :gin
+    t.index ["retrieval_vector"], name: "retrieval_vector_idx", using: :gin
+    t.index ["routine_uses_vector"], name: "routine_uses_vector_idx", using: :gin
+    t.index ["safeguards_vector"], name: "safeguards_vector_idx", using: :gin
+    t.index ["security_vector"], name: "security_vector_idx", using: :gin
+    t.index ["source_vector"], name: "source_vector_idx", using: :gin
+    t.index ["storage_vector"], name: "storage_vector_idx", using: :gin
+    t.index ["summary_vector"], name: "summary_vector_idx", using: :gin
+    t.index ["supplementary_info_vector"], name: "supplementary_info_vector_idx", using: :gin
+    t.index ["system_name_vector"], name: "system_name_vector_idx", using: :gin
+    t.index ["system_number_vector"], name: "system_number_vector_idx", using: :gin
   end
 
 end