This PR is an experiment in database tuning. It adds:
- a bunch of new tsvector columns, for each of our existing sorn content columns
- indexes for those new columns
- triggers to update.vector columns when creating or updating sorns.

It uses our same existing search code, just against the new columns.

#### Results
For single word searches its about three times faster.
For multi word searches, its one second faster.


http://localhost:3000/ with 5890 SORNs -
1. blank search  `Completed 200 OK in 134ms (Views: 65.4ms | ActiveRecord: 67.9ms | Allocations: 36956)`
2. search for ssn `Completed 200 OK in 944ms (Views: 87.4ms | ActiveRecord: 851.3ms | Allocations: 76736)`
3. multiword search for social security number `Completed 200 OK in 2756ms (Views: 52.4ms | ActiveRecord: 1694.2ms | Allocations: 774760)`

**main branch results**
1. blank search - `Completed 200 OK in 123ms (Views: 57.9ms | ActiveRecord: 64.3ms | Allocations: 35703)`
2. search for ssn `Completed 200 OK in 2907ms (Views: 82.8ms | ActiveRecord: 2813.2ms | Allocations: 76911)`
3. multiword search social security number `Completed 200 OK in 3632ms (Views: 121.9ms | ActiveRecord: 2464.4ms | Allocations: 685153)`